Implement U-Boot boot counter support.
With this patch you can alternate between booting the two partitions
two and three. The U-Boot environment itself must be on partition one,
which must be a FAT partition. The partition creation is not currently
automated.

Note that for some reason, even though the fstab is correct, the boot
process refuses to automount the /u-boot directory, so you need to do
this manually to manipulate the environment.

This patch includes the tools needed inside the image (fw_printenv)
and the boot code for U-Boot.