Move port 222 forwarding to separate script.
We should use port 222 forwarding on all hosts, not just proxy hosts,
for the reasons described in the comments of this change. Therefore we
create the new script initialize-user-data.sh, which we should run on
all provisioned hosts. This means one less step to do inside
nested-vm.sh, and we can centralize the steps in the user-data
section, instead of specifying it over and over for every host.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>