Upgrade configuration according to krogoth standard.
This fixes a very elusive bug where you would get a "taskhash
mismatch" error the first time you build using an old config file. The
reason is that poky would auto-upgrade the file behind your back, and
this would cause taskhash calculations to somehow get out of sync. We
would get this error every time in Jenkins because we always start
with a freshly checked out file from this repository.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>