Fix emergency mode error when testing meta-mender:master.
When we build the docker container for the client, we refer to
latest. However, docker-compose will refer to a specific tag,
depending on which branch we're building. In order to avoid having to
look inside the file for this information, or doing a complicated Git
query to try to figure out the Git branch (which is not even
guaranteed to match the tag in the docker-compose file), we force the
tag by passing in a docker-compose file that overrides the first one.

This fixes a problem where the wrong container was used for testing
the client.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>