Fix inability to build certain versions when license has changed.
The issue is a bit complicated: Let's say we have version 1, which has
license A. On the master branch, which comes after version 1, we have
license B. These licenses are checksummed in the bitbake recipes, so
we have to match the version of the source with the recipe.

In the old build script this decision is based on whether or not we
are pushing containers: If we are, the version 1 recipe is chosen, in
order to record a non-SHA version in the binaries. If we are not, then
the git recipe is chosen to get a SHA in the version string, as well
as to be able to build recent versions with updated licenses.

But this does not work if we want to both build the master branch, and
push containers, because it will attempt to build the master version,
that has license B, together with the recipe for version 1, which has
the checksum for license A.

So the fix is this: We assume that whenever we are building from a
branch, it must be a quite recent version, so we use the git
recipe. Only if we are on an exact tag do we choose the versioned
recipes. This will not work correctly for older branches which may
have a different license checksum than the latest git recipe, but this
is a separate problem that can be handled later. We do include the
MENDER_BRANCH and MENDER_ARTIFACT_BRANCH variables in order to make
this future fix easier.

Signed-off-by: Kristian Amlie <kristian.amlie@mender.io>