Make sure that the correct preferred version is set for tags.
This already works correctly if you try to build a tag directly.
However, if building a branch that just happens to be on a tag, the
tag code path is selected even though one probably wanted the non-tag
one. The tag code path should be safe though, as long as we set the
preferred version to the tag name, and not the branch name that was
originally asked for, since this would require a "-git%" suffix in
order to match.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>