Rewrite jenkins build script to support a job matrix.
Many other cleanups as well, such as consolidating duplicate building
and testing steps into generic functions that can be reused for new
boards.

Also take many steps that were previously carried out in the recipe
and add them to the build script so that fixes to them can be tested
in pull requests.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>