Adjust confusing SSH key labels for Kristian.
In the interest of making it easier in the future: let's reveal who's
behind the a10040 label. And also correct the incredibly confusing
kristian@kristian-mender label, which is actually not my key at all,
but the key for accessing the sstate-cache by the build machines,
which just happened to be generated on my machine and got that label.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>