Switch image names from containing MACHINE to containing DEVICE_TYPE.
The main reason for doing this change is that MACHINE is not enough to
separate images from each other when multiple boot loader
configurations are involved. For example vexpress-qemu can be booted
both with U-Boot and with GRUB, and these images are not compatible
with each other. To make sure that images have distinct names, this
change is necessary.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>