First version for the full Mender QA pipeline for GitLab
This commit introduces the first production version of the Mender QA
pipeline for GitLab CI.

Basically, what Jenkins runs through jenkins-yoctobuild-init-script.sh
has been ported to before_script on GitLab pipeline; while the clone of
all microservices that is specified for Jenkins in its configuration, is
done in the main script of GitLab pipeline.

Functionality wise, it reassembles Jenkins builds with the following
exceptions:
* Integration tests are run in parallel with acceptance tests
* There is no support for integration tests with precompiled client
* Upload of artifacts to S3 and Docker pushes (and therefore releases)
are not supported. This is a temporary workaround to not dirturbe the
development process.

The runs are slower than Jenkins due to missing optimizations:
* QA-79: KVM acceleration for qemu emulation
* QA-80: NFS mount for yocto builds

Implementaion detail. Three different Docker images are in use by the
differnt jobs:
* Init stage uses a simple Alpine Linux image
* Build and acceptance tests jobs use a Ubuntu based image, with Docker
as a service (DinD) to build the Docker images of the microservices. A
Linux user is created to run bitbake as non root.
* Integration tests jobs use a privilaged official DinD image to allow
composition of containers and their shared directories as the tests
require.

Changelog: Title

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>