Build at least the x86 client in all scenarios
Historically, the only backend changes will trigger a build with pre
built client for integ tests. In the new GitLab CI, we want to always
built the client to avoid problems when introducing changes.

With this commit, we will force to build x86 even when no BUILD_... is
specified.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>