QA-85: Add backend integration tests job on GitLab CI
Similarly to integration tests, we call `run` script from GitLab
by-passing the yoctobuild-build script. The job shall be run when
RUN_INTEGRATION_TESTS is set, regardless of the client being built or
not (this is a key diference from Jenkins CI).

Modified also the git clone of integration tests to align it to the rest
(i.e. to work for tag revisions)

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>