jenkins-yoctobuild-build: select open or enterprise integration test set
use release_tool to determine what tests to run.

massage arguments for both integration and backend integration tests to
filter out unwanted tests.

this is quick and dirty - pretty sure this bash could be refactored and
structured in a more appealing way, but it does the job and this bit
will be retired anyway in favor of gitlab.

changelog: none

Signed-off-by: Marcin Chalczynski <m.chalczynski@gmail.com>