Reimplement STOP_SLAVE functionality as WAIT_IN_STAGE_INIT|BUILD|TEST.
We can't use the original method, because gitlab has separate stages,
and waiting in the last one is not necessarily what you want (the
environment will be different from preceding stages). Therefore it is
split up into three variables, one for each stage, and each one can be
used to request waiting at the end of that stage.

Also trying to be progressive and stay away from words with negative
historical associations. :-)

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>