Refactor Github PR status updates running in Gitlab CI.
This makes it more robust and less reliant on the enormous
`jenkins-yoctobuild-build` script. The actual posting code is moved to
its own file, and both CI providers call this. Note however, that we
disable updates from within jenkins-yoctobuild-build for Gitlab CI,
and instead rely on its `before_script` and `after_script`
sections. This should make it better at updating the status in the
case of failures.

There are a few weaknesses with this approach:

* Because we can no longer update the `TEST_TRACKER` variable from an
  external script, Jenkins is now unable to post failure updates using
  the shell trap, which means that only pure build and test failures
  will be reported. This was not fullproof before either though. I
  think this is a reasonable tradeoff, since we are moving away from
  Jenkins.

* If a step in the `after_script` fails, then the status will not be
  posted from Gitlab CI. It is possible to fix this by using `||
  ret=$?` style of commands, but it is messy, and will cause extra
  commands to execute when in fact, it should stop. I think this is
  also a reasonable tradeoff, since the `after_script` sections are
  short, much less likely to fail, and it will be visible on Github
  still, just as an unfinished job, rather than a failed one.

Also replaced `env` command with `compgen -A export`, since the former
is not able to handle newlines in variables, of which Gitlab contains
plenty. The latter gives just the names, which is what we want.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>