Remove the single quotes inside SPECIFIC_INTEGRATION_TEST variable
For integration tests, the script expects a test or list of tests (or a
filter, for all that matters) and it will take care of combining it
correctly in the pytest args.

This is different from backend integration tests, where the variable to
control this is PYTEST_ARGS and contains itslef the -k flag for the
filter (and could contain any other flags or options).

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>