Refactor container handling code to support non-mendersoftware links.
First off we switch to using `release_tool.py --list docker` instead
of `git`, since there is no 1-to-1 relationship between our git repos
and Docker images, so this makes it a bit clearer. For the git-only
services we now have a separate section.

Second we use the new `release_tool.py --map-name` to find the correct
Docker URL to push.

Also remove the fallback for when `--list` didn't exist. This was
added in 1.4.x which has not been supported for a long time already.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>