Add GitLab script to trigger client builds publishing artifacts.
The mender client Pipeline is special in the sense that no acceptance
tests (nor full yocto build, for all that matters) is done in the
individual repo. So we will need this script to trigger a part of the
mender-qa Pipeline with publishing logic enable.

Using release_tool from integration repo, construct a string with all
parameters for the curl call(s) specifiying the correct versions for
each repo.

Note that this might lead to more than one pipeline if the client
release branch target is used in more than one Mender integration

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>