Add support to build servers pre-enterprise
At the time, we introduced a new flag to map docker-to-git names in our
release_tool. The build scripts were updated accordingly, but not in a
backwards compatible manner. See commit 2cc785a9.

This commit adds some logic to detect if the release_tool in the
integration version in use has support for such functionality. It also
adds ignore cases for deviceadm as old integration versions will list
it, unfortunately.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>