POC: Generate tests reports with Allure framework
The main feature we want to showcase is having the "history" of failing
tests. This framework supports it in a rather limited way, but let's
give it a try...

Still a PoC that requires manual work to generate the actual reports; to
merge it we need logic to generate the reports from the Pipeline (on
nightly builds) and upload them to an S3 bucket.

Changelog: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>