Make sure non-success status message is posted early to Github.
As more and more steps have been added to the after_script section, it
has become increasingly unlikely that the script will ever get to the
status report, and hence it will not report failure, only pending
forever. This causes a lot of lost waiting time.

Combat this by posting non-success status as soon as we enter the
after_script section. This guarantees that failures will always be
posted, except if the failure lies between this new status update, and
the existing one at the very end. It also guarantees that success
won't be posted too early.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>