Fix script for backend-tests suite selection
It is unsure what the problem is. Debugging on the GitLab CI showed that
the problem in this expresion appears with docker:18-dind and using our
private runner, in the public runner seems to work fine...

The suspected problem is related to the -q flag. Fix that, and also use
grep -e to get the rid of the quoting.

It is unknown wince when this problem is present.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>