Make it possible to omit client and servers builds completely.
This allows the pipeline to go straight to the testing stage, which is
where the build actually happens for many client platforms. This can
save a lot of time. This introduces the `BUILD_CLIENT` and
`BUILD_SERVERS` variables to control the build behavior.

Also organize the variable list a little bit.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>