Fix private repositories not reporting build status in many cases.
The problem is that repo:status permission on Github does not grant
access to the general git API, so we cannot query the commit SHA.
However, we have access to the "status" API. The interesting thing is
that I completely stumbled upon this API, and I cannot find that it's
documented on the the Github API pages. It is not the same as the
"statuses" API, which is documented (and which we use a few lines
later). So it's possible that this will break some day, but until
then...

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>