Remove mender-image-tests handling, use submodule's hash
The previous behavior was counter-intuitive and it is making it
difficult to introduce changes in mender-image-tests while keeping
backwards compatibility with Python 2 (used in warrior).

Remove the special logic and instead rely solely on the Git submodules
hashes. We also have now the aid of dependabot to keep these up to date.

Remove also `--no-pull` flag for acceptance tests, as it will now be
unsupported by the test framework.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>