Switch from "dependencies" to "needs" to enable more parallelism.
Not all "dependencies" keys are changed to "needs". Jobs that execute
strictly last are not changed to avoid having an excessive list of
"needs" entries.

One "dependencies" key was removed because it is in a template, and
the default doesn't make sense on its own.

And finally, since we now "need" jobs that are conditional, we have to
re-specify all the conditions for the jobs that are depended upon.

release_board_artifacts would have a huge list if done this way, so
decided to not use "needs" for this job, since its dependencies
usually finish last anyway.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>