MEN-4420: Support specifying mender-configure version
This is an independent component, so it won't be used in integration
tests and generally we will just built with "latest" (iow, yocto will
decide based on available recipes).

But for the cases where we might need some manual tweaking, this logic
will allow us to build any desired version.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>