[test_full_integration] Clean-up pip packages install
Namely: remove docker-compose from apk install, so that we instead
install the docker-compose version specified in requirements.txt file.

Indirectly, this is hiding the "cryptography upgrade" issue, as the OS
package installs an older version of it and the subsequent call to pip
install will ignore this dependency.

Removed also the now obsolete pip/pip3 calls, as both are Python 3 in
today's Alpine Linux.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>