QA-255: Only build the mender client when BUILD_CLIENT is true
This splits the integration_test_jobs up into two separate entities. One with a
needs: build_client, and one without, which will not run the build_client job.

A few changes were needed:
* The integration tests will only run when RUN_INTEGRATION_TESTS = true

And then, depending on the value of BUILD_CLIENT, the integration test job will
run with our without a client build.

The client version chosen, when the client is not built from source is set
in the integration version chosen already, so no configuration needed.

Changelog: None
Signed-off-by: Ole Petter <ole.orhagen@northern.tech>