Introduce special variable for Docker building jobs.
Previously we relied on a messy combination of ONLY_BUILD and testing
for TEST_QEMUX86_64_UEFI_GRUB. This would break if you manually
started a job where BUILD was set and TEST wasn't.

The new variable BUILD_DOCKER_IMAGES is much more descriptive of what
will happen if it is true. We enable this in the docker-building job
only. I chose to keep ONLY_BUILD as well, since these two decisions
are, strictly speaking, independent, even though they are both enabled
in the same job.

Signed-off-by: Kristian Amlie <kristian.amlie@northern.tech>