Restore the tarball as input for the yocto build.
The meta-mender recipes are meant to use the tarball as the input, not
the uncompressed files. This is the way that users will built it, so we
need to verify the same workflow.

Partially reverts: ""Monitoring: decompress the tar archive and use it
for build, repo name fix"

This partially reverts commit 2c0c70a3efa67db111168de1c94d11a39970987a.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>