Remove source checkout of monitor-client and related check
The release_tool check is from the Jenkins times, when the runner's
workspace was shared and could run into some bad race conditions.

By removing the check, we can also remove the extra (unnecessary)
checkout of the monitor-client to make it explicit that we only use the
tarball as input for the yocto build.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>