QA-322: Rework child pipelines triggers
Have two explicit trigger stages: packages (mender-dist-packages) and
images (mender-convert). These are controlled by explicit variables
BUILD_* TEST_* and implicit variable PUBLISH_RELEASE_AUTOMATIC.

This should now align the full release process with the child pipelines,
and give the release coordinator the required flexibility to trigger
"build and publish" or "build and test" kind of pipelines.

Additionally, mender-convert version to use can be manually set by the
release coordinator with MENDER_CONVERT_REV - in the future we will have
it as part of the release_tool.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>