[FIX] scripts: update-image fail early and format output
A failure occured where one of the `gcloud` image queries returned two
images and broke the commands using this output.
This commit sets the '-e' shell flag to fail early if an error should
occur, in addition to fixing the specific case that made the script fail
on this insident.
changelog: none

Signed-off-by: Alf-Rune Siqveland <alf.rune@northern.tech>