yocto related jobs: Refactor script to error on missing artifacts
Move all code of the template to `before_script` so that the jobs can
have in their script the collection of the stage artifacts, so that if
something errors there the build job errors (`after_script` is allowed
to fail).

The main motivation is to catch these issue where the last qemu image is
not collected in `build:client:qemu` but the problem is only seen once
we are running integration tests...

Once in the area, the syntax for an if snippet has been updated.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>