Pipeline: introduce new stage yocto:build-n-test
Moving there the "build" job for QEMU image used by integ tests and the
"build and test" jobs for client acceptance tests.

The motivation is to clearly have requirements for the yocto biuld, like
mender-monitor package (and soon mender-gateway package) so that they
can also depend on init and share the same workspace.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>