build.yml: Hardcode `ref: master` when including .gitlab-ci-build-package.yml
This was introduced to be able to specifiy the version to include, but
this turned out to be a problem for our tooling, as it will break when
trying to include a "pull/XXX/head" kind of revision on PR builds.

Remove it and rely on master. We expect this build job to be relatively
stable so using master should be fine - for special cases we can always
hardcode here a different revision.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>