fix(yocto-build-and-test.sh): Download only mender-binary-delta tarball
For the new versions, we will need the `licenses` directory which is
only included in the tarball.

Modify the script to get only the tarball (extract the generator) and
later either use the tarball as-is for kirkstone or newer or extract it
for older (which better represents what users do in these cases).

Changelog: None
Ticket: None

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>