chore: Set empty dependencies for ci jobs in .post stage
This shortcut should speed up the status report jobs.

The default behavior is to get all dependencies from all previous
stages, which for this last stage translates to almost 4 minutes of
downloading everything.

We cannot use `needs` as we still want it to run at the end of the
pipeline.

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>