ci: Build mender-configure from external source
Historically, we have treated the repo as an "independent component",
although it is really checked out together with the rest and we have git
recipes in meta-mender.

This commit aligns the component with the rest.

The hidden motivation of this change is to workaround a failure in the
client acceptance tests, where building a non-latest mender-configure
client and then building a full image fails with an error related to
sstate. It is suspected that the bug (raise condition?) will also be
triggered for any of the other components if they weren't to use
external source.

It is possible to reproduce locally, using today's scarthgap, with
something like:
```
python3 -m pytest -k '(test_preferred_versions and mender-configure and 1.1.1) or (test_order and 2)'
```

See the following CI output:
* https://gitlab.com/Northern.tech/Mender/mender-qa/-/jobs/7123097490

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>