ci: Remove dependencies for generation CI job
This job has no real dependencies. The actual trigger depends both on
the generation and the QEMU build, so by setting up no dependencies in
the generation in effect we can start the integration tests in parallel
of the client acceptance tests (as it used to be in the old times).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>