ci: Conditionally set the components versions for Yocto builds
Amends commit c8580ed.

Besides conditionally setting the external source, we also need to
conditionally _override_ the version only when the components are being
checked out by the build system (i.e. when they are part of the
release).

Otherwise for example releasing Mender Client 5.0.x is in effect always
using master of the independent components instead of the specified
latest available tag (for example for `mender-artifact`).

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>