ci: Remove build and publish of `mender-cli` tool
This tool needs to be moved out of the legacy release process.

So far we kept it in `mender-qa` pipeline while we work on its own
release process but recent changes in the source repository breaks the
pipeline.

Removing it prematurely to not block releases of the Mender Client or
other tools that still live here.

Ticket: MEN-8269

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>