ci: Remove `release-tool` from repositories initialization
Use instead (for now, at least) a hardcoded list of repositories. Since
we only have Mender Client + yocto repos the list is manageable.

The magic keyword "latest" will skip the repository checkout. This is
part of the decoupling design between Mender Client, Mender Gateway and
tools (more on this on the commits to come).

This commit by itself should not alter the build.

Ticket: QA-1166

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>