build: Rework `build:docker` job
So that it does not depend on `init:workspace` and instead checks out
directly the repo that contains the build code (which is `integration`
today and we are planning to move to somewhere else).

The real motivation behind this commit is isolating the dependency of
`integration` repo to this job, opening up for further work to fully
remove `release-tool` usage from the pipeline.

Ticket: QA-1166

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>