ci: download mender-orchestrator master tarball
This is the missing piece of for test_mender_orchestrator to run. Since
this is a closed-source component, we need to download and add the
tarball.

For now we hardcode the version to `master` (see comment in yocto-build-and-test.sh).
The plan is to align this with the other closed-source components (e.g. mender-gateway) in
QA-1180

Signed-off-by: Daniel Skinstad Drabitzius <daniel.drabitzius@northern.tech>