ci: Replace source for building the Virtual Device
The Virtual Device for application updates is now owned by the new
repository `mender-client-subcomponents`.

I think this is the official RIP for `integration` repo source code in
the context of `mender-qa` :) Its version is still used as per to
represent the Mender Client "bundle" and to version the integration
tests themselves (although this we will address soon too).

Ticket: MEN-9015

Signed-off-by: Lluis Campos <lluis.campos@northern.tech>